refactor: 카테고리 API 호출 로직을 CategoryClient로 분리
