+++ src/main/java/shop/wannab/frontservice/couponpolicy/CouponPageDataDto.java
@@ -3,6 +3,7 @@ package shop.wannab.frontservice.couponpolicy;
 import java.util.List;
 import lombok.Getter;
 import lombok.Setter;
+import shop.wannab.frontservice.category.controller.response.CategoryHierarchyDto;
 
 @Getter
 @Setter
